feat: JCEF w/ CEF 130.1.9+gfc42567+chromium-130.0.6723.70 (#9)
* Fix CefDisplayHandler.onFullscreenModeChange name (see #239)

Method names should follow Java naming conventions.

* Add two flags to CefPdfPrintSettings

* CefApp.getInstance: Support switch values containing '=' (fixes #470)

* Update to CEF 126.2.0+g5c56e98+chromium-126.0.6478.62

* Update to CEF 127.3.1+g6cbb30e+chromium-127.0.6533.100

Disabled signal handlers on POSIX systems and removed the previous signal
handler fix.

* Fix Java signature for CefRequestHandler#onRenderProcessTerminated method

* Update to CEF 130.1.9+gfc42567+chromium-130.0.6723.70

This adapts JCEF to a significant change in CEF: the removal of the Alloy Bootstrap. It makes JCEF use the Alloy Runtime mode for "normal" browser windows (because Chrome-style windows can't be integrated into existing native parents, which is how JCEF integrates browsers into Java UI) and changes the DevTools handling to open the DevTools in a separate pop-up window. The latter is necessary because DevTools are unsupported in an Alloy-style window by CEF at the moment, thus it's no longer possible to integrate the DevTools into Java windows.

* fix: pack_loading_disabled on context

* fix: no AWT OSR

---------

Co-authored-by: Vladislav Rassokhin <vladrassokhin@gmail.com>
Co-authored-by: Sander van den Berg <svandenberg@uplandsoftware.com>
Co-authored-by: Rene Schneider <slartie@posteo.de>
Co-authored-by: Loïc Frasse-Mathon <scooby-lolo97@hotmail.fr>
Co-authored-by: David Cernoch <dcernoch@uplandsoftware.com>